attached_assets/requirements.txt
fastapi
uvicorn
httpx
requests
pydantic
python-dotenv
sqlalchemy
databases
